requirements.txt
time
opencv-python
numpy
keyboard
serial
tk
Pillow
DateTime
sympy